The MrFOR_simrep Module

MrFOR_simrep — Module

MrFOR_simrep

Simulation settings repository for MrFOR models.

This package sole function is load_settings.

The settings are stored in this repository as yaml files in:

[project]

scenarios

[scenario]

Where the default project and the default scenarios ae both named "default".

Module Index

MrFOR_simrep.load_settings

Detailed API

MrFOR_simrep.load_settings — Function

load_settings(project="default",scenario="default")

Load the settings for a specific project and scenario. The settings are returned as a dictionary.